saveAuthData.mjs
Creating function to set token and user data to local storage and edit the firebaseAuth.mjs to do not replace the token in the local storage.
Trying to create the feature to display username, hide login btn and display logout btn. Needs improvement because after page reload must be logged in to display it again